+++ Sources/Models/AuthMethod.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public enum AuthMethod {
+    case endpoint(authEndpoint: String)
+    case authRequestBuilder(authRequestBuilder: AuthRequestBuilderProtocol)
+    case authorizer(authorizer: Authorizer)
+    case inline(secret: String)
+    case noMethod
+}
