+++ Sources/Models/EventHandler.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct EventHandler {
+    let id: String
+    let callback: (PusherEvent) -> Void
+}
