+++ Sources/Models/PusherChannel.swift
@@ -1,29 +1,5 @@
 import Foundation
 
-public enum PusherChannelType {
-    case `private`
-    case presence
-    case normal
-
-    public init(name: String) {
-        self = Self.type(forName: name)
-    }
-
-    public static func type(forName name: String) -> PusherChannelType {
-        if name.components(separatedBy: "-")[0] == Constants.ChannelTypes.presence {
-            return .presence
-        } else if name.components(separatedBy: "-")[0] == Constants.ChannelTypes.private {
-            return .private
-        } else {
-            return .normal
-        }
-    }
-
-    public static func isPresenceChannel(name: String) -> Bool {
-        return PusherChannelType(name: name) == .presence
-    }
-}
-
 @objcMembers
 open class PusherChannel: NSObject {
     // Access via queue for thread safety if user binds/unbinds events to a channel off the main queue
@@ -201,14 +177,4 @@ open class PusherChannel: NSObject {
             unsentEvents.insert(QueuedClientEvent(name: eventName, data: data), at: 0)
         }
     }
-}
-
-public struct EventHandler {
-    let id: String
-    let callback: (PusherEvent) -> Void
-}
-
-public struct QueuedClientEvent {
-    public let name: String
-    public let data: Any
 }
