+++ Sources/Models/PusherClientOptions.swift
@@ -1,25 +1,5 @@
 import Foundation
 
-public enum PusherHost {
-    case host(String)
-    case cluster(String)
-
-    public var stringValue: String {
-        switch self {
-        case .host(let host): return host
-        case .cluster(let cluster): return "ws-\(cluster).pusher.com"
-        }
-    }
-}
-
-public enum AuthMethod {
-    case endpoint(authEndpoint: String)
-    case authRequestBuilder(authRequestBuilder: AuthRequestBuilderProtocol)
-    case authorizer(authorizer: Authorizer)
-    case inline(secret: String)
-    case noMethod
-}
-
 @objcMembers
 @objc public class PusherClientOptions: NSObject {
     public var authMethod: AuthMethod
