+++ Sources/Models/PusherEventError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum PusherEventError: Error {
+
+    case invalidFormat
+    case invalidDecryptionKey
+    case invalidEncryptedData
+}
