+++ Sources/Models/PusherPresenceChannel.swift
@@ -176,14 +176,3 @@ public typealias PusherUserInfoObject = [String: AnyObject]
         return findMember(userId: id)
     }
 }
-
-@objcMembers
-@objc public class PusherPresenceChannelMember: NSObject {
-    public let userId: String
-    public let userInfo: Any?
-
-    public init(userId: String, userInfo: Any? = nil) {
-        self.userId = userId
-        self.userInfo = userInfo
-    }
-}
