+++ Sources/Models/QueuedClientEvent.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct QueuedClientEvent {
+    public let name: String
+    public let data: Any
+}
