+++ Sources/Services/PusherConnection.swift
@@ -989,53 +989,3 @@ extension PusherConnection: PusherEventQueueDelegate {
         group.wait()
     }
 }
-
-internal struct PusherAuthError: Error {
-    enum Kind {
-        case notConnected
-        case noMethod
-        case couldNotBuildRequest
-        case invalidAuthResponse
-        case requestFailure
-    }
-
-    let kind: Kind
-
-    var message: String?
-
-    var response: URLResponse?
-    var data: String?
-    var error: NSError?
-}
-
-@objc public class PusherAuth: NSObject {
-    public let auth: String
-    public let channelData: String?
-    public let sharedSecret: String?
-
-    public init(auth: String, channelData: String? = nil, sharedSecret: String? = nil) {
-        self.auth = auth
-        self.channelData = channelData
-        self.sharedSecret = sharedSecret
-    }
-}
-
-@objc public enum ConnectionState: Int {
-    case connecting
-    case connected
-    case disconnecting
-    case disconnected
-    case reconnecting
-
-    static let connectionStates = [
-        connecting: "connecting",
-        connected: "connected",
-        disconnecting: "disconnecting",
-        disconnected: "disconnected",
-        reconnecting: "reconnecting"
-    ]
-
-    public func stringValue() -> String {
-        return ConnectionState.connectionStates[self]!
-    }
-}
